move formatting logic into SemanticDateUtil
